Server/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Cache
redis==5.0.1

# HTTP Client (for Perenual API)
httpx==0.26.0

# Task Queue
celery==5.3.6

# Image Processing
pillow==10.2.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
